Enable FEC for system, vendor, odm and product partitions
Forward error correction (FEC) is an error correction technique to
detect and correct a limited number of errors in corrupted data.
The following test checks, the number of encoding data blocks on the
FEC device and the number of generator roots.
If one of these numbers is zero, the test fails.

Test: vts-tradefed run vts -m VtsSecurityAvb -t \
VtsSecurityAvb#AvbTest.SystemDescriptor_64bit

Jira-Id: ADM-4121
Change-Id: I892e9d9fa728e5bc1b08b14f8cbfc7e1b8a7fa10
Signed-off-by: Leonid Komaryanskiy <leonid.komaryanskiy@globallogic.com>