Add boottime_prop permission for vendor init
This commit adds get_prop permisson that is needed by init.{hardware}.rc
scripts. This permission allows us to mark end of coldboot stage - the
ro.boottime.init.cold_boot_wait is set at the final moment of this stage.
This is the earliest moment for any binderized HAL start.

Jira-ID: ADM-4166
Signed-off-by: Dmytro Firsov <dmytro.firsov@globallogic.com>
Change-Id: Ia1ed67acc84dcabbcdf8cddde45483efb5017d83