Disable camera service by default for all boards
For now all Renesas boards do not support camera feature by default.
It means that all of the boards do not have `CameraProvider` or
`ExternalCameraProvider` services on the HAL layer. But at the same
time image contains `Camera` and `CameraProxy` services on the
framework layer which are redundant in the current configuration.

It is proposed to reduce using camera components as much as possible:
 - the `Camera` service can be excluded from the image and build
   process at all;
 - start of the `CameraProxy` service can be skipped during
   the `SystemService` initialization.
The patch also contains the granting of SELinux rights for
`SystemServer` and `Gallery2` to read vendor properties. The current
`aosp/master` branch has a special property `exported_camera_prop` for
this purpose, so in the next Android version the granted rights should
be revised.

Additional changes:
 - remove the `USE_CAMERA_STUB` definition from the build script
   because it does not have effect for Renesas boards.

Some additional information can be found here:
 - $ANDROID_BUILD_TOP/frameworks/base/services/java/com/android/\
                      server/SystemServer.java

Test: the output for the following commands should be empty:
      # lshal | grep -i cam
      # logcat -d SystemServerTiming:* *:S | grep -i cam
      # service list | grep -i cam
Jira-Id: ADM-4216, ADM-4424
Change-Id: Iab369b2235572ca67f4f3c579b0d3f6f8ce46dbc
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>