DeviceCommon.mk: disable device_admin feature
The `device_admin` feature was added into the Renesas configuration a
long time ago when automotive configuration did not exist yet. Enabling
this feature turns on some CTS tests which are not applicable for car
boards. As far as it is not mentioned in the current "car product"
configuration it is proposed to disable it.

Test: run cts -m CtsDevicePolicyManagerTestCases
Jira-Id: ADM-4486, ADM-4209, ADM-4220
Change-Id: I118c1cbc8e7c78b0789adf9c5119fa93d9cfffa6
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>