implemented custom low memory killer for ION region
As the AOSP lmk can't track ION memory usage, it leads to UI
freezing in case if some of ION consumers has a memory leak.
This service controls ION memory usage. If OOM has been occurred or
the threshold has been crossed, then the ion_lmk finds a non-banned
process, which consumes the largest amount of memory and sends
SIGKILL to it.

Currently, banned process list consists of:
1. "system_server"
2. "surfaceflinger"
3. "android.systemui"
4. "android.hardwar"
5. "car.carlauncher"
These services are critical, so they shoudln't be tracked.
The list is printed on service start to logcat.

The threshold for remaining free ION memory can be set by
'-t' argument in MBytes:
"ion_lmk -t 100" - will set threshold to 100MB

Jira-Id: ADM-4622
Change-Id: Ie8b70bc3dd8ccfad935eadb918e580199ae69738
Signed-off-by: Leonid Komaryanskiy <leonid.komaryanskiy@globallogic.com>
Signed-off-by: Vitaliy Vasylskyy <vitaliy.vasylskyy@globallogic.com>