Remove config_UsbDeviceConnectionHandling_component to fix endless boot
UsbHostManagementActivity from UsbDeviceConnectionHandling component
sometimes is launched and become idle earlier than system finish
booting. This leads to race condition in ActivityStackSupervisor which
prevent system boot up completely. When this property is not set
UsbHostManagementActivity is not launched on start up and issue is not
reproduced.

Test: Manual.
1. Boot up DUT and set any secure lock. Example:
adb shell locksettings set-pin 0000
2. Attach a usb device.
3. Reboot the DUT several times and verify observed.

JiraId: ADM-4500
Signed-off-by: Artem Radchenko <artem.radchenko@globallogic.com>
Change-Id: I57a360b96611544d9e25741570d34475061cd724