move EVS application json configs into its folder
These is just an auxiliary change that intended to make dependencies
more convenient.

The config json files are used by the EVS application at run-time. They
are stored in the EVS HAL folder due to historical reasons. When the
AOSP version of the EVS application used to use it was a proper way to
have platform-specific configuration files. But now when we have our own
fork of the EVS application it would be more correct to store them in
the EVS application project folder.

Note: this changes must be applied simultaneously for the following
three projects:
 - $ANDROID_BUILD_TOP/vendor/renesas/hal/evs;
 - $ANDROID_BUILD_TOP/vendor/renesas/apps/evs;
 - $ANDROID_BUILD_TOP/device/renesas/common.

Test: build, flash a dut and ensure that EVS works well.
Jira-Id: ADM-5092
Change-Id: Ib461a8f6555a35a9aa16a7fb0f088bc0ae5f213e
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>