Update EAS enable/disable commands
/proc/sys/kernel/sched_energy_aware:
===================

Enables/disables Energy Aware Scheduling (EAS). EAS starts
automatically on platforms where it can run (that is,
platforms with asymmetric CPU topologies and having an Energy
Model available). If your platform happens to meet the
requirements for EAS but you do not want to use it, change
this value to 0.

Signed-off-by: Dmytro Prokopchuk <dmytro.prokopchuk@globallogic.com>
Jira-ID: ADM-5124
Change-Id: I844a58b7e2a12a417960b83729be008a5c660dcd